Move TrackSelectionListFeature and its related files to the list folder
