+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/track_selection/delegate/TrackSelectionListDelegate.kt
@@ -10,7 +10,7 @@ import org.hyperskill.app.android.R
 import org.hyperskill.app.android.core.view.ui.adapter.decoration.itemDecoration
 import org.hyperskill.app.android.track_selection.adapter.TrackAdapterDelegate
 import org.hyperskill.app.android.track_selection.model.TrackSelectionRecyclerItem
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature
 import ru.nobird.android.ui.adapterdelegates.dsl.adapterDelegate
 import ru.nobird.android.ui.adapters.DefaultDelegateAdapter
