+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/injection/AppGraph.kt
@@ -59,7 +59,7 @@ import org.hyperskill.app.topics_to_discover_next.injection.TopicsToDiscoverNext
 import org.hyperskill.app.topics_to_discover_next.injection.TopicsToDiscoverNextDataComponent
 import org.hyperskill.app.track.injection.TrackComponent
 import org.hyperskill.app.track.injection.TrackDataComponent
-import org.hyperskill.app.track_selection.injection.TrackSelectionListComponent
+import org.hyperskill.app.track_selection.list.injection.TrackSelectionListComponent
 import org.hyperskill.app.user_storage.injection.UserStorageComponent
 
 interface AppGraph {
