+++ shared/src/commonMain/kotlin/org/hyperskill/app/track_selection/list/injection/TrackSelectionListComponent.kt
@@ -1,0 +1,10 @@
+package org.hyperskill.app.track_selection.list.injection
+
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.Action
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.Message
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.ViewState
+import ru.nobird.app.presentation.redux.feature.Feature
+
+interface TrackSelectionListComponent {
+    val trackSelectionListFeature: Feature<ViewState, Message, Action>
+}
