+++ shared/src/commonMain/kotlin/org/hyperskill/app/track_selection/list/injection/TrackSelectionListComponentImpl.kt
@@ -1,10 +1,10 @@
-package org.hyperskill.app.track_selection.injection
+package org.hyperskill.app.track_selection.list.injection
 
 import org.hyperskill.app.core.injection.AppGraph
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature.Action
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature.Message
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature.ViewState
-import org.hyperskill.app.track_selection.view.TrackSelectionListViewStateMapper
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.Action
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.Message
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.ViewState
+import org.hyperskill.app.track_selection.list.view.TrackSelectionListViewStateMapper
 import ru.nobird.app.presentation.redux.feature.Feature
 
 class TrackSelectionListComponentImpl(
