+++ shared/src/commonMain/kotlin/org/hyperskill/app/track_selection/list/injection/TrackSelectionListFeatureBuilder.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.track_selection.injection
+package org.hyperskill.app.track_selection.list.injection
 
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
@@ -8,12 +8,12 @@ import org.hyperskill.app.progresses.domain.interactor.ProgressesInteractor
 import org.hyperskill.app.sentry.domain.interactor.SentryInteractor
 import org.hyperskill.app.study_plan.domain.repository.CurrentStudyPlanStateRepository
 import org.hyperskill.app.track.domain.interactor.TrackInteractor
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListActionDispatcher
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature.Action
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListFeature.Message
-import org.hyperskill.app.track_selection.presentation.TrackSelectionListReducer
-import org.hyperskill.app.track_selection.view.TrackSelectionListViewStateMapper
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListActionDispatcher
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.Action
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListFeature.Message
+import org.hyperskill.app.track_selection.list.presentation.TrackSelectionListReducer
+import org.hyperskill.app.track_selection.list.view.TrackSelectionListViewStateMapper
 import ru.nobird.app.presentation.redux.dispatcher.wrapWithActionDispatcher
 import ru.nobird.app.presentation.redux.feature.Feature
 import ru.nobird.app.presentation.redux.feature.ReduxFeature
